MVP of action-based permission checks

+++ service/src/main/kotlin/fi/espoo/evaka/backupcare/BackupCareController.kt
@@ -8,12 +8,12 @@ import fi.espoo.evaka.Audit
 import fi.espoo.evaka.shared.BackupCareId
 import fi.espoo.evaka.shared.DaycareId
 import fi.espoo.evaka.shared.GroupId
-import fi.espoo.evaka.shared.auth.AccessControlList
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
-import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.db.mapPSQLException
 import fi.espoo.evaka.shared.domain.FiniteDateRange
+import fi.espoo.evaka.shared.security.AccessControl
+import fi.espoo.evaka.shared.security.Action
 import org.jdbi.v3.core.JdbiException
 import org.springframework.format.annotation.DateTimeFormat
 import org.springframework.http.ResponseEntity
@@ -28,7 +28,7 @@ import java.time.LocalDate
 import java.util.UUID
 
 @RestController
-class BackupCareController(private val acl: AccessControlList) {
+class BackupCareController(private val accessControl: AccessControl) {
     @GetMapping("/children/{childId}/backup-cares")
     fun getForChild(
         db: Database.Connection,
@@ -36,7 +36,7 @@ class BackupCareController(private val acl: AccessControlList) {
         @PathVariable("childId") childId: UUID
     ): ResponseEntity<ChildBackupCaresResponse> {
         Audit.ChildBackupCareRead.log(targetId = childId)
-        acl.getRolesForChild(user, childId).requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR, UserRole.FINANCE_ADMIN, UserRole.STAFF, UserRole.SPECIAL_EDUCATION_TEACHER)
+        accessControl.requirePermissionFor(user, Action.Child.READ_BACKUP_CARE, childId)
         val backupCares = db.read { it.getBackupCaresForChild(childId) }
         return ResponseEntity.ok(ChildBackupCaresResponse(backupCares))
     }
@@ -49,7 +49,7 @@ class BackupCareController(private val acl: AccessControlList) {
         @RequestBody body: NewBackupCare
     ): ResponseEntity<BackupCareCreateResponse> {
         Audit.ChildBackupCareCreate.log(targetId = childId, objectId = body.unitId)
-        acl.getRolesForChild(user, childId).requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Child.CREATE_BACKUP_CARE, childId)
         try {
             val id = db.transaction { it.createBackupCare(childId, body) }
             return ResponseEntity.ok(BackupCareCreateResponse(id))
@@ -66,7 +66,7 @@ class BackupCareController(private val acl: AccessControlList) {
         @RequestBody body: BackupCareUpdateRequest
     ): ResponseEntity<Unit> {
         Audit.BackupCareUpdate.log(targetId = backupCareId, objectId = body.groupId)
-        acl.getRolesForBackupCare(user, backupCareId).requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.BackupCare.UPDATE, backupCareId)
         try {
             db.transaction { it.updateBackupCare(backupCareId, body.period, body.groupId) }
             return ResponseEntity.noContent().build()
@@ -82,7 +82,7 @@ class BackupCareController(private val acl: AccessControlList) {
         @PathVariable("id") backupCareId: BackupCareId
     ): ResponseEntity<Unit> {
         Audit.BackupCareDelete.log(targetId = backupCareId)
-        acl.getRolesForBackupCare(user, backupCareId).requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.BackupCare.DELETE, backupCareId)
         db.transaction { it.deleteBackupCare(backupCareId) }
         return ResponseEntity.noContent().build()
     }
@@ -96,8 +96,7 @@ class BackupCareController(private val acl: AccessControlList) {
         @RequestParam("endDate") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) endDate: LocalDate
     ): ResponseEntity<UnitBackupCaresResponse> {
         Audit.DaycareBackupCareRead.log(targetId = daycareId)
-        acl.getRolesForUnit(user, daycareId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN, UserRole.UNIT_SUPERVISOR, UserRole.STAFF)
+        accessControl.requirePermissionFor(user, Action.Unit.READ_BACKUP_CARE, daycareId)
         val backupCares = db.read { it.getBackupCaresForDaycare(daycareId, FiniteDateRange(startDate, endDate)) }
         return ResponseEntity.ok(UnitBackupCaresResponse(backupCares))
     }
